--- a/pkg/distributor/distributor.go
+++ b/pkg/distributor/distributor.go
@@ -2133,8 +2133,8 @@ func (d *Distributor) deduplicateActiveSeries(ctx context.Context, matchers []*l
 	}
 
 	maxResponseSize := math.MaxInt
+	// For native histograms we are going to return metrics level information so we don't need to limit the response size.
 	if !nativeHistograms {
-		// We are going to return metrics level information, this limit doesn't make much sense in that case
 		if limit := d.limits.ActiveSeriesResultsMaxSizeBytes(tenantID); limit > 0 {
 			maxResponseSize = limit
 		}