… CHANGELOG.md
… cmd/mimir/config-descriptor.json
… docs/sources/mimir/configure/configuration-parameters/index.md
… integration/query_frontend_active_series_test.go
… pkg/frontend/querymiddleware/roundtrip_test.go
… pkg/frontend/querymiddleware/shard_active_series.go
… pkg/frontend/querymiddleware/shard_by_series_base.go
… pkg/querier/cardinality_analysis_handler.go
… pkg/querier/cardinality_analysis_handler_test.go
… pkg/querier/distributor_queryable_test.go
… pkg/querier/querier_test.go
… pkg/util/validation/limits.go