@krajorama 2024-05-03
Benchmark that additional uint64 in data structure for deduplication does nothing
```
goos: linux
goarch: amd64
pkg: github.com/grafana/mimir/pkg/querier
cpu: AMD Ryzen 7 4700U with Radeon Graphics         
                                │   old.txt   │            new.txt            │
                                │   sec/op    │   sec/op     vs base          │
ActiveSeriesHandler_ServeHTTP-8   258.7µ ± 3%   253.8µ ± 4%  ~ (p=0.075 n=10)

                                │   old.txt    │            new.txt             │
                                │     B/op     │     B/op      vs base          │
ActiveSeriesHandler_ServeHTTP-8   146.4Ki ± 0%   146.3Ki ± 0%  ~ (p=0.353 n=10)

                                │  old.txt   │           new.txt            │
                                │ allocs/op  │ allocs/op   vs base          │
ActiveSeriesHandler_ServeHTTP-8   115.0 ± 0%   115.0 ± 0%  ~ (p=1.000 n=10)
```